CDAP-3605 Entity deletion should remove its metadata.
Address comments and rename InMemoryBusinessMetadataStore to NoOpBusinessMetadataStore.

Change for rename InMemoryBusinessMetadataStore to NoOpBusinessMetadataStore.

Change name from DistributedBusinessMetadataStore to DefaultBusinessMetadataStore since it is used in standalone mode.
